Update patch set 1
Patch Set 1:

> Jason: PTAL. Do we have a strong reason not to perform the check for ARCVM? I'm guessing we don't.

It was because:
- ADB on VM does not depend on networking,
- Android IP of VM might not always be 100.115.92.6 (esp if we turn on multinet)

Is this patch created to have a consolidated arc.go for VM and container? If so, I would suggest removing `waitNetworking()` because as of now ARC container also connects to ADB independent of networking.

Patch-set: 1
Reviewer: Gerrit User 1346383 <1346383@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>