Update patch set 2
Patch Set 2:

(11 comments)

Note that we need to wait for EventGenerator (CL:2220655) to be included in the uprev'd chrome before submitting this. That CL is included in 85.0.4159.0 and you could check goldeneye to see chrome version goes to/beyond that version.

Patch-set: 2
Reviewer: Gerrit User 1000101 <1000101@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>