Update patch set 9
Patch Set 9:

(1 comment)

Awesome, thanks for doing this. One more thing (sorry for not bringing this up before), is there a way to detect whether or not CSA is supported? Ideally, we'd like to expect only the channel switch event if we support it, and only the disconnect event if we don't.

Patch-set: 9
Reviewer: Gerrit User 1243287 <1243287@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>