Rename firmware.CheckBootMode to firmware.BootMode
When I originally named firmware.CheckBootMode, my intention for the
test was to exercise checking which boot mode the DUT is in.
But, you cannot check which boot mode the DUT is in without first
rebooting it into the different boot modes.
So, this test exercises both switching and checking the DUT boot mode.
Hence, the test should have a more general name to reflect that it
covers all the boot-mode functionality.

BUG=b:155425293
TEST=firmware.BootMode

Change-Id: Icc5e7fb79955b62c4955e7f6501af35a132b2b0f