Update patch set 4
Patch Set 4: Code-Review+2

(2 comments)

LGTM with a nit. Giving CR+2, but please wait for a while so that lacros folks have chances to have a look (the Cc-ed folks are in Tokyo, so they should be able to see this in Sunday of our timezone).

Patch-set: 4
Reviewer: Gerrit User 1000479 <1000479@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2