KernelCrash: Remove .bios_log file if it exists.
Previously, since we couldn't assume the .bios_log file existed, we
didn't check for it and hence didn't clean it up, since we only cleaned
up files we checked for.

Resolve this by adding the bios_log file(s) to the list of files to be
cleaned up.

BUG=None
TEST=tast run $DUT  crash.KernelCrash.*; ls /var/spool/crash on DUT

Change-Id: I72186aed7030bb9c8943a6104cfab73f5e8688d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2222716
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>