Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+2
> 
> in few tests we have request.vmEnabled (ureadahed, gmscore service). It is good to deprecate this as well, might be good to do in separate CL.

Thanks. Could you let me know which CL the proto(?) was introduced? (I'm not familiar with that)

> Wonder if we can get rid of vm suffix.

To do that, we need to have a dep like "android_p_or_r" which I'm not sure if we want to introduce. We also have to merge the VMBooted() precondition to the container one for many tests. Let me discuss this separately.

Patch-set: 2
Reviewer: Gerrit User 1000303 <1000303@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>