Look for .log file in WatchdogCrash as well.
The .log file is actually always generated after a watchdog crash (with
the first 100 lines of dmesg), so look for it in WaitForCrashes and
remove it when the test is done.

NB: I don't have a device that runs the watchdog available to test on.
see https://tastboard.corp.google.com/test/crash.WatchdogCrash for a
list of such devices.

BUG=chromium:1039381
TEST=fast_build.sh

Change-Id: I5f5bcb9e7876137e364b2182b0fbee132cdd4138