Look for .log file in WatchdogCrash as well.
The .log file is actually always generated after a watchdog crash (with
the first 100 lines of dmesg), so look for it in WaitForCrashes and
remove it when the test is done.

NB: I don't have a device that runs the watchdog available to test on.
see https://tastboard.corp.google.com/test/crash.WatchdogCrash for a
list of such devices.

BUG=chromium:1039381
TEST=fast_build.sh

Change-Id: I5f5bcb9e7876137e364b2182b0fbee132cdd4138
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2222894
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Auto-Submit: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>