Servo: Add call option to avoid unpacking response
Sometimes, it's helpful to avoid unpacking the XMLRPC response from
servo. The attached bug is an example of this. So is the modified
SetString code, which was a workaround for an unwanted response.

This CL adds an option to servo.call which causes call.run() to skip
unpacking the call's XMLRPC response, and updates SetString (and
SetActChgPort) to use that new option.

BUG=b:157915707
TEST=example.ServoEcho, plus a call to SetString

Change-Id: Ibcb64d79fe0c2216d5db70b3f0a78857555b1d9a