tast-tests/video: PlaybackPerf: verify decoder before measurements
ToT carries out all intended measurements before verifying
that the decoder used was the expected one (hw, sw, etc).
This CL changes that to letting the video roll for one
second (because chrome:media-internals takes time to be
populated), then veryfing the decoder, and only then starting
to collect measurements.

BUG=chromium:1071841
TEST=tast run DUT video.PlaybackPerf.* on kohaku

Change-Id: Ia17fb8f97c880069e6795af1976cd7b91904edb4