tast-tests: Declare dependencies for security.Microcode
Skip the test when running on a VM, and only run on amd64 platforms.
The latter is to be replaced eventually with a HardwareDep that
restricts to the SoC vendor Intel (or whatever subset of platforms the
test can verify once the SoC vendor HardwareDep becomes available).

BUG=chromium:1090362
TEST=Test runs on eve but gets skipped on betty.

Cq-Depend: chromium:2228210
Change-Id: I40812b8315e3aa596ba8beb22b57073557d912a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2228959
Commit-Queue: Mattias Nissler <mnissler@chromium.org>
Tested-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>