Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (1 comment)
> 
> > Patch Set 3:
> > 
> > (1 comment)
> > 
> > How about instead of all these, install a DLC, remove the verified pref, restart dlcservice and check the install. doesn't that give you the same thing rather than trying to yet maintain another dlcservice implementation detail here? :)
> 
> Yes, that would be another way to do it. This method here is already short as is though 😊
> It's also inline with how cros deploy works + tests dlc internals.
> 
> Current:
>  - Copy to Cache
>  - Call Install
> 
> Other:
>  - Install
>  - Delete verified pref
>  - Restart + Install

Up to you, but keep in mind that your current approach is trying to mimic a lot of internal implementation details of dlcservice like file permission, where the files locate, etc that one of us need to maintain whenever we change internals of dlcservice. This is a functional testing and it should not do that. Although, removing the verified flag is also an implementation detail, it is much less intrusive than setting file permissions.  
https://www.simform.com/unit-testing-vs-functional-testing/

And as always, shorter solution, mostly, but not always the better one 😊

cheers,

Patch-set: 3
Reviewer: Gerrit User 1163331 <1163331@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>