Update patch set 14
Patch Set 14: -Verified

> Patch Set 14: Verified-1 Code-Review-1
> 
> (1 comment)
> 
> Hey Matthew, do you mind keeping the 16GB setting?

I'm looking into doing this through the installer UI.  Does it have to be 16GB, or is it ok if I just set it to the maximum allowed by the installer?  As the UI uses a slider it will be difficult to get an exact disk size.

Patch-set: 14
Reviewer: Gerrit User 1370239 <1370239@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0
Label: Verified=0