Retry Open when creating syslog Readers
When opening a system log, the log may be in the middle of being
rotated, so there is a race condition if we call os.Open() once and then
immediately fail the test if it fails. This is particularly true of
NewChromeReader, because Chrome is often restarted at the end of a test
(from defer cr.Close and the like), so the Chrome log is rotated right
at the moment a new test starts up. Example:
https://stainless.corp.google.com/browse/chromeos-autotest-results/swarming-4c929a6566f1a010/autoserv_test/tast/results/tests/ui.ChromeCrashNotLoggedIn.broker_breakpad_mock_consent/

Try opening the log more than once before failing the test.

BUG=None
TEST=tast run ... ui.ChromeCrashNotLoggedIn.broker_*_mock_consent

Change-Id: Ie2a1c727eb2d3faf65cb09c86c9676541cd2b167