Update patch set 4
Patch Set 4:

I'm having trouble getting platform.MemoryPressureRecorder to run on my DUT, but everything else passes, so I guess it's OK?

Alexis, the test times out here:
2020/06/04 15:49:17 [15:49:16.812] Creating new page with URL https://plus.google.com/discover
2020/06/04 15:49:17 [15:49:16.818] Connecting to Chrome target 3D835E2D370A743E281F5DB6EA7C36A5

And there's a popup on the DUT screen saying "Sign-in has changed", and a notification that there was a sign-in error.

Do you know what could be causing this?

The error is reported properly as a context timeout, so at least that part of my change is working.

Patch-set: 4