Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> IMO the computation of sanitized username is not something suitable at integration test level. It doesn't test the behavior but the implementation already. I don't think this is something we wanna achieve at this level. And, did you check if this is tested by any unittest already?
> Also, lots of tests already implies its correctness, no? If it only increases the runtime of the testsuite w/o providing any extra information but just noise to the log, please reconsider the legitimacy of this test item.


The reason why I'm testing the computation implementation of the GetSanitizedUsername() is because once a user logs in, the sanitized username will stick as the folder name on disk. That means we will not be able to change the implementation of GetSanitizedUsername() without a carefully planned migration so user doesn't get screwed (losing their homedir because its name changed) when the implementation changes.

In that case, I aim to test the implementation of GetSanitizedUSername() in order to catch any implementation change that required migration. Or in another word, the exact implementation of GetSanitizedUsername() is actually part of the "behaviour" because users expect them to be consistent throughout the entire lifetime of their chromebook (sans powerwash).

Patch-set: 2
Reviewer: Gerrit User 1314760 <1314760@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>