Use NEW_READY signal to detect LibAssistant is ready
This signal was introduced a while ago, and is sent
out when LibAssistant signals it is ready to handle
queries.

Now that LibAssistant has been upreved to 1.46s, this
signal is reliable and can be used in the Tast tests
to know when LibAssistant is ready.

BUG=b:142428671
TEST=Tast tests with filter "*ssistant*"

Change-Id: Ic335701ac30631ca7ef213be96e498b575ca5fba
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2230333
Tested-by: Jeroen Dhollander <jeroendh@google.com>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Auto-Submit: Jeroen Dhollander <jeroendh@google.com>
Reviewed-by: Greg Edelston <gredelston@google.com>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>