Set a short sleep between mouse drag moves.
We observed the checkWindowsNum after the second mouse drag was failed.
The first mouse drag might not be fully handled yet when the second
mouse drag started, which might cause the second drag to fail to drag
the window. Thus we set a sleep between two drags, and also wrap
checkWindowsNum into testing.Poll for more robustness.

BUG=chromium:1091535
TEST=tast run IP ui.DragTabInClamshellPerf
Change-Id: Ibb4d67ed32154617c34c1feec00f0c038ef679d0