firmware: In boot to Rec, set V4Role & defer reset
This replicates Autotest functionality needed for switching DUT boot
mode to Recovery.

When the Servo USB will be used for recovery boot, the DUT needs to set
the servo_v4_role to 'snk'.
Then when Servo is closed, it sets the role back to 'src'.
In order to do that, this CL also adds Servo.Defer() to cause a call to
wait until Servo.Close().

BUG=b:158298086
TEST=manual: On a DUT with Servo_v4 Type C, run boot to Recovery; verify
that v4 role is 'snk'; call svo.Close(); verify that v4 role is 'src'.

Change-Id: If8de248377206317a7b6938a171ce5707f313711