Update patch set 5
Patch Set 5: Code-Review+2

(2 comments)

Other food for thought: should we put an arbitrary bound on the values, so that this reports a failure if scans are way outside the bounds we think they should be? Can be a follow-up, if you'd like.

Patch-set: 5
CC: Gerrit User 1158184 <1158184@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1307891 <1307891@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1132414 <1132414@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2