minidump: Avoid using context.Background.
When chrome.New hits the timeout on restarting the ui job, it captures
minidumps of various TPM-related daemons to help debugging slow TPM
operations.

Since this logic is triggered on timeout, the minidump package used to
run external commands to capture minidumps with the background context.
This was not a good use of the background context because context
deadline is essentially ignored while capturing minidumps.

This change fixes the minidump package to use the given context. To
allow the capturing to work, we have to pass an alive context to the
minidump package. This is done by limiting the timeout to restart the
ui job (uiRestartTimeout) to the value less than the whole login
timeout (LoginTimeout, gaiaLoginTimeout). Actually the current values
where uiRestartTimeout is longer than LoginTimeout makes little sense;
essentially uiRestartTimeout is ignored today.

BUG=chromium:1087058
TEST=fast_build.sh -T
TEST=tast run DUT ui.ChromeLogin

Change-Id: I826b41b36b6e4936c430f5ab807fe6d650d80f0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2233437
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>