minidump: Avoid using context.Background.
When chrome.New hits the timeout on restarting the ui job, it captures
minidumps of various TPM-related daemons to help debugging slow TPM
operations.

Since this logic is triggered on timeout, the minidump package used to
run external commands to capture minidumps with the background context.
This was not a good use of the background context because context
deadline is essentially ignore while capturing minidumps.

This change fixes the minidump package to use the given context. To
allow the logic to work, we have to pass an alive context to the
minidump package. This is possible by limiting the timeout to restart
the ui job, i.e. uiRestartTimeout.

uiRestartTimeout is currently longer than LoginTimeout, but actually
it makes little sense because uiRestartTimeout is essentially ignored.

BUG=chromium:1087058
TEST=fast_build.sh -T
TEST=tast run DUT ui.ChromeLogin

Change-Id: I826b41b36b6e4936c430f5ab807fe6d650d80f0c