wifi.SecChange: Enable packet capturing in the test.
The test almost always fails on grunt DUTs with failure in getting IP
from DHCP server but DHCP server does not seem to receive any requests.
With leased grunt DUTs, the test always passes and the problem cannot
be reproduced. Let's enable packet capturing to gather more information
about the failure so that we can narrow down the problem.

BUG=b:156569968
TEST=`tast run $DUT wifi.SecChange` and make sure pcap is generated.

Change-Id: I28fd7f455aef8be9d5def77b53349772e78717c0