wifi.HostapHwsim: robust ProhibitedTechnologies handling
If origProhibited contains "wifi" already, we set it a second time, and
Shill doesn't like that. We need to smartly determine whether to append
"wifi" or not.

This is all much easier if we have a slice to handle, rather than a
comma-separated string. Let's split/join such a slice for our
intermediate handling.

BUG=none
TEST=run it, with "wifi" in the prohibited list before the test
     e.g., `dbus-send --system --print-reply \
	      --dest=org.chromium.flimflam / \
	      org.chromium.flimflam.Manager.SetProperty \
	      string:ProhibitedTechnologies string:"wifi"

Change-Id: I7459a8c2de953eaf3d57623df706dd9688559303
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2798565
Tested-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>
Reviewed-by: Yen-lin Lai <yenlinlai@chromium.org>