Update patch set 3
Patch Set 3:

> What do you folks think? Would it be better to have this checked in, or not?

I slightly prefer having this merged after the router can support it, as the lab is quite busy now. But this seems to fail fast (should fail when spawning hostapd) and the overhead of a Tast test is much smaller than an Autotest test, so I don't have strong opinion.

Not sure if I missed some bigger downside of deferring this. Dean@, WDYT?

Patch-set: 3