Update patch set 3
Patch Set 3:

(1 comment)

> Patch Set 3:
> 
> (1 comment)
> 
> Perhaps it would be more idiomatic to wait for the goroutine completion in StopRecording using a channel instead of using a Mutex.

There is already such a channel (`recordingStatus`). StopRecording waits until it receives a value on that channel, indicating goroutine completion.

Actually, the mutex is not signaling goroutine completion here. It is used to ensure that the goroutine (i.e., its context) is not cancelled while a Snapshot is in place. We don't want to have a half-finished snapshot.

Patch-set: 3