Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Hi,
> > 
> > Indeed it looks very stable, though our basic guideline asks consistent passes for 20 builds (see https://chrome-internal.googlesource.com/chromeos/chromeos-admin/+/master/doc/tast_add_test.md#step-4_remove-attribute-to-make-it-run-on-infra), and the history for this test is a bit shorter than this criteria.
> > 
> > Can you wait for this?
> 
> Thanks Jun, I will wait for this.
> However, it would be nice to catch potential issues prior to landing CLs as the tests are very stable atm, but while development is actively being made.
> 20 builds seem kinda long in my opinion~

Yeah indeed. I also wish it can be a bit shorter. But to make that shorter, I'll have to discuss with other tast owners. Can you elaborate your situation a bit more about why this informational attr should be removed earlier?

Note: if a new feature is landed and breaks the test because the expectation is changed, the test would fail until the test fix is made -- I don't think that breaks '20 consecutive builds' guideline. See the footnote of the linked document, saying

> if a test fails for a few runs for an unrelated and fixed issue, you might be able to ignore those failures on assessing stability. Use your best judgement to expand test coverage while keeping CQ stable.

Patch-set: 1