Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> > Patch Set 1:
> > 
> > My only concern with this solution is that it assumes the cups-clear-state job is either running or completed by the time tast finishes logging in. If that assumption is ever false, the test will start failing again.
> 
> From the discussion in the bug, it sounds like we feel safe with the current assumption.

I see the concern as well, but from my testing this does not seem to be a large risk.

I ran the tests specified in the commit message 30 times and did not observe any failures. There appears to be ~5 seconds between the two cups-clear-state runs (the first being triggered by logging in, the second by the printer.ResetCups() call). This seems like a large enough buffer to minimize my concerns.

Patch-set: 1