Update patch set 1
Patch Set 1:

> Also, is it okay not to handle the failure case of assistant.Disable() in deferred call?

If you look at most instances of a (Disable, Close, Cleanup, etc) method, they will simply return an error. In the case that a call to that method is deferred, the error is ignored. This is ok because the test has already succeeded at that point. Cleanup issues generally should not count towards a test failure(personally, I really think that Tast should have specially error types for setup and cleanup that are reported differently, but that is a separate debate). By just returning an error, it lets the caller choose how to handle the error.

Technically, it would be better if the callers logged the errors instead of just ignoring them, but that is pretty uncommon in Tast(probably mostly for brevity sake). So really the change should be at the location of the defer to:

defer func() {
  if err := someCleanup(); err != nil{
    s.Log("Failed to do some cleanup: ", err)
  }
}()

Patch-set: 1