Add the workaround to prevent ui.OverviewPerf failures
As is seen in the comment and noted in the bug, it turns out that
the overview animation for a browser window will be canceled when
the status bubble for the browser window is visible.

Although this test does nothing with the mouse, previous test may
use a mouse, and the last location may fall into a location which
causes the status bubble to be visible, and we can't prevent this
systematically.

Ideally this should be fixed in chromium side, since I believe
the overview animation shouldn't be canceled by the status bubble,
but for the time being, the workaround of moving mouse location
would be a good thing to do.

BUG=chromium:1091547
TEST=tast run ui.FilesAppZipPerf ui.OverviewPerf on sentry

Change-Id: I4f10cd62c06b3f85b6fcb1af6b3d7339c80d2efd