Update patch set 14
Patch Set 14:

> Patch Set 14: -Verified
> 
> > Patch Set 14: Verified-1 Code-Review-1
> > 
> > (1 comment)
> > 
> > Hey Matthew, do you mind keeping the 16GB setting?
> 
> I'm looking into doing this through the installer UI.  Does it have to be 16GB, or is it ok if I just set it to the maximum allowed by the installer?  As the UI uses a slider it will be difficult to get an exact disk size.

I don't understand the slider UI issue. Can we get a Chrome command line option?
Overall the needs are such that we can live with anything larger than 16GB as long as there is some space left on stateful for tast to operate. Say at least 5GB for tast and 16GB for VM for a typical 32GB device that has 21GB stateful partition.

Notice that the devices that we are targeting have anything from 21GB to 200+GB space on stateful. We just need some programmatic control to split this across all these devices and have some guarantees on the split going forward in time.

Patch-set: 14