Update patch set 1
Patch Set 1:

Any progress on this CL?

> If we can't store the converted image file and must keep converting, then I think it is fair to do it fast. Yes this will increase thermal issues and maybe it is going to be cancelled by having to wait for idle/cool CPU afterward, but that is a separate problem. I am fine with the change.

We currently set up benchmark mode before extracting the video, so we won't wait for the CPU to become idle afterwards. But Maybe it's not a bad idea to do so, the wait time should be limited. Benchmark mode also disables thermal throttling, maybe we should only do that after extracting the video?

Patch-set: 1