Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Also, is it okay not to handle the failure case of assistant.Disable() in deferred call?
> 
> If you look at most instances of a (Disable, Close, Cleanup, etc) method, they will simply return an error. In the case that a call to that method is deferred, the error is ignored. This is ok because the test has already succeeded at that point. Cleanup issues generally should not count towards a test failure(personally, I really think that Tast should have specially error types for setup and cleanup that are reported differently, but that is a separate debate). By just returning an error, it lets the caller choose how to handle the error.
> 
> Technically, it would be better if the callers logged the errors instead of just ignoring them, but that is pretty uncommon in Tast(probably mostly for brevity sake). So really the change should be at the location of the defer to:
> 
> defer func() {
>   if err := someCleanup(); err != nil{
>     s.Log("Failed to do some cleanup: ", err)
>   }
> }()

The reason why I suggest failing the test because of teardown / cleanup error is because the error implies the DUT's condition is not restored in some sort and might affect subsequent tests. I've deal with several inter-test flakiness and would like to see a better mechanism to facilitate debugging.

Why logging doesn't work: the subsequent failed test's owner does not have an idea what makes the DUT behave differently. It is unlikely to look at every precedent tests' log as long as they pass the test. Also, for the owner of the test, as long as it passes the test, the owner is unlikely to exam every log to see if the cleanup/teardown fails.

Why test fail works: it gives a stronger hint that the failed test's following tests might be affected. Also, the failed test's owner would have to see why the cleanup doesn't work as its test fails from time to time.

Patch-set: 1