Exclude Chromeboxes/Chromebits from UI perftests
Right now, presentation callbacks are not coming when physical
displays don't exist, and the chromeboxes / chromebits in the
lab are not actually connected to displays. That will cause
test failures on UI perftests which rely on them.

Right now, input latency (presentation time) metrics rely on that,
but some of animation smoothness metrics based on throughput will
also rely on that.

For the sanity of the test data, we want to exclude all of such
devices from UI perftests.

BUG=chromium:1036108
TEST=tast run '("group:crosbolt" && "name:ui.*")' on ninja

Change-Id: Ib7e91bfd9337ebe07967e20bf1a5fb887f933eff