Update patch set 2
Patch Set 2:

(3 comments)

Thanks for doing this. Having some comments but they would be minor.

BTW, I'm now wondering like, if it is possible to bind multiple sub-actions to a node. For example, I think most of the cases, the steps would be like "from root, find one UI item, make action, wait for the next UI appears", that would be written as:
  uig.Steps(uig.Root().FindWithTimeout(ui.FindParams{...}).LeftClick(),
            uig.Root().FindWithTimeout(ui.FindParams{...}),
            ...)

But always starting from Root would be cumbersome. Maybe uig.FromNode() would be fine, but something like:
  uig.Root().Steps(uig.FindWithTimeout(ui.FindParams{...}).LeftClick(), uig.FindWithTimeout(...))

is possible, it would be kinda happy to me. That is just a comment, you don't have to pursue this, but if you have some ideas, that would be much appreciated.

Patch-set: 2