tast-tests/video: Add test that draws a video on a canvas.
This CL adds a new test case that draws a video onto a 2D canvas and
checks four pixels for the correct color. The motivation is that this
path is easy to break in Chrome and is used to draw video thumbnails in
the UI (for example, in the camera capture app).

As case in point, this path broke after switching to the new video
decoder and was not detected by our current tests. The current failure
is that the video is not drawn at all for some resolutions (like 360p
and 1080p). Another plausible failure mode is getting artifacts on the
edges of the canvas (e.g., a green line). Therefore, it should be
sufficient to check only the four corners of the canvas for the correct
color after drawing the video.

To avoid worrying about when the video is drawn to the canvas, the test
videos are generated from a single image (so it doesn't matter when the
video is drawn, we'll always get the same image). Each video is simply 4
rectangles of different colors. They were generated by creating an image
in GIMP and then using the following ffmpeg command:

  ffmpeg -loop 1 \
         -i <src-image>.png \
	 -c:v libx264 \
	 -pix_fmt yuv420p \
	 -t 1 -profile:v baseline \
	 still-colors-<resolution>.mp4

Bug: 1092806
Test: ran on eve to ensure it passes/fails as expected.
Change-Id: Ie894a3bb35b92c7b59f3312d9e89f70e5cb9de8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2242603
Tested-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>