WIP - tast-tests: Use ui.Node in a11yTree
Use ui.Node instead of custom automation struct. This meant that
the way the diff of a11yTree was computed needed to be re-done.
In this change, after obtaining the rootNode of the application,
we will check whether each child matching the parameters of the expected
tree exists.

By defining the expected tree as ui.Node, this also allows for the
removal of the JSON files.

BUG=b:155949540
TEST=tast run DUT_IP arc.AccessibilityTree

Change-Id: I06716a7025ea1c09b297916c7c4d39d96b40feef