tast-tests/webrtc: wait for 1st frame decode in RTCPeerConnectionPerf
RTCPeerConnectionPerf high-load test cases such as vp8_hw_multi_vp9_4x4
take a substantial amount of time to start decoding "remote" video
frames. This is identified as a "black frame" when it should be logged
as fewer overall decode frames (the former is a test failure whereas
the latter is a performance-tracked issue). This CL simply waits until
the first frame is decoded before analysing its contents.

BUG=b:158848650
TEST=tast run DUT webrtc.RTCPeerConnectionPerf.vp8_hw_multi_vp9_4x4
  on kohaku, celes

Change-Id: I0e1c681d9b3635bd8448a3ab9a72bf18390e5275