Don't run WatchdogCrash on bob, scarlet, or kevin.
Per the linked bug these boards have unstable watchdog drivers. The
boards are old enough and the flakes rare enough that the firmware team
is not going to fix them. We still want the test to be stable and do not
want to break DUTs in the lab / cause them to fall into a NEEDS_REPAIR
state, so skip the test on these boards.

BUG=chromium:1069618
TEST=fast_build.sh -T  # I don't have a DUT with a watchdog to test on.

Change-Id: Id2e6779ac2b436ceefa12a3646a0935b740179b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2247203
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Jun Mukai <mukai@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>