Update patch set 1
Patch Set 1:

But you don't think the test is flaky? In which case disabling the test to push a release is the wrong approach, no? We can't release to users while ToT is broken so it doesn't actually let us release any sooner. The right thing to do is figure out what's wrong and fix it.
If you have some change not in the VM uprev blocked by this then let's roll back the VM test pin uprev so you can get your change in.

If you really need some code pushed now, then CrOS oncall would be the people to ask about chumping.

sidereal@ is taking a look at the failure, so I'd sync with him on current status.

Patch-set: 1