Remove hard-coded string check for What's New title
Checking that the title of the page is present is a good way to see if
something on the page has loaded and is visible, but it makes the test
quite brittle. Another option is to create a Chrome connection to the
What's New PWA and check for document.readyState property to be
complete, indicating that the page has finished loading. Ideally, unique
IDs could also be provided for some elements on the page and we could
check for their existence as well.

BUG=chromium:1077091
TEST=tast run localhost:2222 ui.WhatsNewSmoke

Change-Id: Id63ebd399ae8fc786191b126ed008601cc545c34