Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> @mukai for +2
> https://tastboard.corp.google.com/test/platform.DLCServicePreloading

One failure is seen on hatch at 13280, failed with "Install failed: failed to install: exit status 70" Do you know what was the cause?
As is described in the footnote of https://chrome-internal.googlesource.com/chromeos/chromeos-admin/+/master/doc/tast_add_test.md, 
>  if a test fails for an unexplained reason (e.g. kernel crash not observed in other tests), you should not remove the informational attribute. On the other hand, if a test fails for a few runs for an unrelated and fixed issue, you might be able to ignore those failures on assessing stability

So it is okay if:
- you know what was the cause, and the fix was already made
- this was known as some infra issues or unrelated to this test

Otherwise, I'd like to wait for a bit. Sorry.

Patch-set: 1