tast-tests: ui.WindowCyclePerf tracks initial metrics separately
ui::AnimationThroughputReporter start tracking under when animation
starts, which includes the initial time of gpu in preparing for
the animation before the first frame is processed. This ends up
reporting a lower smoothness for WindowCycleView (50% -> 20%). This
is because old code path based on AnimationMetricsRecorder considers
only number of committed frames hence excludes the initial time.

This CL records the initial metrics separately so that it is still
on track and regressions could be caught. Meanwhile, the smoothness
metrics would not be considered as regressed and triggers alerts.

BUG=chromium:1021774
TEST=tast run IP ui.WindowCyclePerf

Change-Id: Ia94a3d5c0a9198a7137d72b9bc7b13278b50d6f7