network: VPNConnect: Wait for ethernet state.
After restarting shill the ethernet service could be down while the test
is running. Adding a check for the ethernet service state to be online
after starting shill.

Also, adding a wait for the connection to be established after
configuring the static IP address.

BUG=b:157677857
TEST=tast run $DUT network.VPNConnect.l2tp_ipsec_psk

Change-Id: Ia6930cf25da5815ec2a227dd75001c42deec040d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2248906
Tested-by: Arowa Suliman <arowa@chromium.org>
Commit-Queue: Arowa Suliman <arowa@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Oleh Lamzin <lamzin@google.com>
Reviewed-by: Brian Norris <briannorris@chromium.org>