tast-test: Check packets loss percentage for ping.
The Autotest ping runner has a default threshold of 20% for the packets
loss and the ping command will return an error if the threshold is
exceeded.

Unlike the autotest the caller compare the res.Loss with the
ping.DefaultLossThreshold (=20%) or another desired perecentage.

BUG=b:159175087, b:159139389
TEST=run tast tests
wifi.PowerSave                    [ PASS ]
network.VPNConnect.l2tp_ipsec_psk [ PASS ]

Change-Id: If6fc08da268e9906f8384e9dde3233933b207953