video: Add test to check the contents of a video.
This CL creates a test that:

1) Starts playing a video.
2) Switches to full screen by injecting the 'f' key.
3) Takes a screenshot using the CLI tool.
4) Analyzes the screenshot to check the four corners of the video.

To analyze the screenshot, we first rotate it if necessary (this is
needed on, e.g., kukui where the screen is rotated). Then, we locate the
top and bottom of the video to exclude the black strips on top and
bottom that are not part of the video. Finally, we sample the color of
the four corners of the video and compare against hard coded
expectations. Note that the test videos consist of still images, so it
doesn't matter exactly when the screen is captured. However, we wait a
few seconds before taking the screenshot in order to exclude the "Press
Esc to exit full screen" message from the screenshot.

This test is able to verify the screenshot regardless of whether or not
HW overlays were used (because the screenshot tool captures them). Also,
as far as I could tell, the night light mode does not affect the output
of the screenshot tool.

Bug: b:159158467
Test: passes on eve and kukui.
Change-Id: If0938f320114bb7f2143e16f9d47be325b943910