Update patch set 14
Patch Set 14:

In fact, tunneled1x doesn't need to install the client cred and in that case clientCred would be empty, but we still needs the result of ShillServiceProperties (and it checks the Pin and ClientSlotID fields). So check if the clientCred is empty before installing.

Another solution is to check if clientCred is empty in ShillServiceProperties (pass if empty), and then implement a NeedsTPMStore returning false in tunneled1x.Config. This will not even set up a TPMStore, so should be faster.

Though I just uploaded the first solution, now I think that the second one is better :p

Patch-set: 14