Update patch set 2
Patch Set 2:

TBH, can't really sign off this CL because this CL chain just propagates the wrong design pattern even to other modules outside hwsec test, including (but not limited to) everything that is realted to/dependent on CmdRunner.

zuan@:
Please switch to gPRC before we provide the feature to other teams. The design pattern we have now is not scalable.

Patch-set: 2