Update patch set 1
Patch Set 1:

(1 comment)

cc: oshima since this is rather related to the purpose of the test;

I am wondering if we really need youtube here. I believe PiP is working on any <video> element. If so, I guess you could use a simple html file with some testing video data. And there are plenty of them as for testing playback feature of video.
If that's okay, use of <video> is much much simpler and better for testing. The data files are local to tests, which is also beneficial.

Patch-set: 1
CC: Gerrit User 1001380 <1001380@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>