Update patch set 1
Patch Set 1:

It doesn't have to be a youtube, but the UI on top of the video can also affect overlay and performance so it's important to measure with that element. We can simulate what youtube does, but IF using youtube isn't problematic, then youtube is fine with me.

A separate note. We should measure both Chrome PIP and Android PIP. You can dot it in a separate CL though.

Patch-set: 1