tast-test: Add wifi.DuplicateBSSID
The Tast test is migrated from network_WiFi_MaskedBSSID.py, an Autotest.
Changed the name of the test to a positive form DuplicateBSSID, which
describes the behaviour of the test more clearly.

In addition, instead of running a scan to make sure both SSIDs appear in
the scan, the test make sure that we can  connect to both SSIDs.

BUG=b:158696331
TEST=run the test wifi.MaskedBSSID

Change-Id: I4b90a6b82b27d63db6ba6136706d2aa10076f0f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2250839
Tested-by: Arowa Suliman <arowa@chromium.org>
Commit-Queue: Arowa Suliman <arowa@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Yen-lin Lai <yenlinlai@google.com>